pyproject.toml
[project]
name = "AdvPI"
version = "1.0.0"
description = "Official implementation of 'Investigating Robustness in Vision-Language Models via Adversarial Prompt Illumination'"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "datasets==4.8.4",
    "fastparquet==2026.3.0",
    "gdown==5.2.1",
    "hydra-core==1.3.2",
    "kagglehub==1.0.0",
    "lpips==0.1.4",
    "matplotlib==3.10.8",
    "nltk==3.9.3",
    "numpy==2.2.6",
    "open-clip-torch==3.3.0",
    "pandas==2.3.3",
    "pyarrow==23.0.1",
    "pydicom==3.0.2",
    "pyvoro==1.3.2",
    "scapy==2.7.0",
    "scikit-learn==1.7.2",
    "scipy==1.15.3",
    "sentence-transformers==5.1.2",
    "tokenizers==0.22.2",
    "torch==2.9.1",
    "transformers==4.57.3",
    "typing-extensions==4.15.0",
    "wget==3.2",
]
